+++ src/python/dxpy/cli/parsers.py
@@ -320,8 +320,25 @@ class PrintInstanceTypeHelp(argparse.Action):
 instance_type_arg = argparse.ArgumentParser(add_help=False)
 instance_type_arg.add_argument('--instance-type',
                                metavar='INSTANCE_TYPE_OR_MAPPING',
-                               help=fill('Specify instance type(s) for jobs this executable will run; see --instance-type-help for more details', width_adjustment=-24),
+                               help=fill('''When running an app or applet, the mapping lists executable's entry points or "*" as keys, and instance types to use for these entry points as values.  
+When running a workflow, the specified instance types can be prefixed by a stage name or stage index followed by "=" to apply to a specific stage, or apply to all workflow stages without such prefix. 
+The instance type corresponding to the "*" key is applied to all entry points not explicitly mentioned in the --instance-type mapping. Specifying a single instance type is equivalent to using it for all entry points, so "--instance-type mem1_ssd1_v2_x2" is same as "--instance-type '{"*":"mem1_ssd1_v2_x2"}'. 
+Note that "dx run" calls within the execution subtree may override the values specified at the root of the execution tree.
+See dx run --instance-type-help for details.
+''', width_adjustment=-24, replace_whitespace=False),
                                action='append').completer = InstanceTypesCompleter()
+
+instance_type_arg.add_argument('--instance-type-by-executable',
+                               metavar='DOUBLE_MAPPING',
+                               help=fill(
+                                   '''Specifies instance types by app or applet id, then by entry point within the executable.
+The order of priority for this specification is:
+  * --instance-type, systemRequirements and stageSystemRequirements specified at runtime
+  * stage's systemRequirements, systemRequirements supplied to /app/new and /applet/new at workflow/app/applet build time
+  * systemRequirementsByExecutable specified in downstream executions (if any)
+See dx run --instance-type-help for details.
+''', width_adjustment=-24, replace_whitespace=False))
+
 instance_type_arg.add_argument('--instance-type-help',
                                nargs=0,
                                help=fill('Print help for specifying instance types'),
@@ -383,6 +400,17 @@ def process_instance_type_arg(args, for_workflow=False):
         else:
             # is a string
             args.instance_type = _parse_inst_type(args.instance_type)
+
+def process_instance_type_by_executable_arg(args):
+    if args.instance_type_by_executable:
+        if args.instance_type_by_executable.strip().startswith('{'):
+            # expects a map, e.g of entry point to instance type or instance count
+            try:
+                args.instance_type_by_executable = json.loads(args.instance_type_by_executable)
+            except ValueError:
+                raise DXParserError('Error while parsing JSON value for --instance-type-by-executable')
+        else:
+            raise DXParserError('Value given for --instance-type-by-executable could not be parsed as JSON')
 
 def process_instance_count_arg(args):
     if args.instance_count: